MacroModelling.get_IRF — Function

See get_irf

MacroModelling.get_SS — Function

See get_steady_state

MacroModelling.get_first_order_solution — Function

See get_solution

MacroModelling.get_irf — Method

get_irf(
    𝓂,
    parameters;
    periods,
    variables,
    shocks,
    negative_shock,
    initial_state,
    levels
)

Return impulse response functions (IRFs) of the model in a 3-dimensional array. Function to use when differentiating IRFs with repect to parameters.

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

parameters [Type: Vector]: Parameter values in alphabetical order (sorted by parameter name).

Keyword Arguments

periods [Default: 40, Type: Int]: number of periods for which to calculate the IRFs

variables [Default: :all]: variables for which to calculate the IRFs. Inputs can be either a Symbol (e.g. :y or :all), Tuple{Symbol, Vararg{Symbol}}, Matrix{Symbol} or Vector{Symbol}. Any variables not part of the model will trigger a warning.

shocks [Default: :all]: shocks for which to calculate the IRFs. Inputs can be either a Symbol (e.g. :y, :simulate, :none, or :all), Tuple{Symbol, Vararg{Symbol}}, Matrix{Symbol} or Vector{Symbol}. :simulate triggers random draws of all shocks. Any shocks not part of the model will trigger a warning. :none in combination with an initial_state can be used for deterministic simulations.

negative_shock [Default: false, Type: Bool]: calculate a negative shock. Relevant for generalised IRFs.

generalised_irf [Default: false, Type: Bool]: calculate generalised IRFs. Relevant for nonlinear solutions. Reference steady state for deviations is the stochastic steady state.

initial_state [Default: [0.0], Type: Vector{Float64}]: provide state from which to start IRFs. Relevant for normal IRFs.

levels [Default: false, Type: Bool]: return levels or absolute deviations from steady state

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

get_irf(RBC, RBC.parameter_values)
# output
4×40×1 Array{Float64, 3}:
[:, :, 1] =
 0.00674687  0.00729773  0.00715114  0.00687615  …  0.00146962   0.00140619
 0.0620937   0.0718322   0.0712153   0.0686381      0.0146789    0.0140453
 0.0688406   0.0182781   0.00797091  0.0057232      0.00111425   0.00106615
 0.01        0.002       0.0004      8.0e-5         2.74878e-29  5.49756e-30

MacroModelling.get_irf — Method

get_irf(
    𝓂;
    periods,
    algorithm,
    parameters,
    variables,
    shocks,
    negative_shock,
    generalised_irf,
    initial_state,
    levels
)

Return impulse response functions (IRFs) of the model in a 3-dimensional KeyedArray

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

Keyword Arguments

periods [Default: 40, Type: Int]: number of periods for which to calculate the IRFs

algorithm [Default: :first_order, Type: Symbol]: solution algorithm for which to show the IRFs

parameters : If nothing is provided, the solution is calculated for the parameters defined previously. If a vector with parameter values, or a named tuple is provided and the parameters differ from the previously defined the solution will be recalculated.

variables [Default: :all]: variables for which to calculate the IRFs. Inputs can be either a Symbol (e.g. :y or :all), Tuple{Symbol, Vararg{Symbol}}, Matrix{Symbol} or Vector{Symbol}. Any variables not part of the model will trigger a warning.

shocks [Default: :all]: shocks for which to calculate the IRFs. Inputs can be either a Symbol (e.g. :y, :simulate, :none, or :all), Tuple{Symbol, Vararg{Symbol}}, Matrix{Symbol} or Vector{Symbol}. :simulate triggers random draws of all shocks. Any shocks not part of the model will trigger a warning. :none in combination with an initial_state can be used for deterministic simulations.

negative_shock [Default: false, Type: Bool]: calculate a negative shock. Relevant for generalised IRFs.

generalised_irf [Default: false, Type: Bool]: calculate generalised IRFs. Relevant for nonlinear solutions. Reference steady state for deviations is the stochastic steady state.

initial_state [Default: [0.0], Type: Vector{Float64}]: provide state from which to start IRFs. Relevant for normal IRFs.

levels [Default: false, Type: Bool]: return levels or absolute deviations from steady state

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

get_irf(RBC)
# output
3-dimensional KeyedArray(NamedDimsArray(...)) with keys:
↓   Variables ∈ 4-element Vector{Symbol}
→   Period ∈ 40-element UnitRange{Int64}
◪   Shock ∈ 1-element Vector{Symbol}
And data, 4×40×1 Array{Float64, 3}:
[:, :, 1] ~ (:, :, :eps_z):
        (1)           (2)           …  (39)            (40)
  (:c)    0.00674687    0.00729773        0.00146962      0.00140619
  (:k)    0.0620937     0.0718322         0.0146789       0.0140453
  (:q)    0.0688406     0.0182781         0.00111425      0.00106615
  (:z)    0.01          0.002             2.74878e-29     5.49756e-30

MacroModelling.get_irfs — Function

See get_irf

MacroModelling.get_moments — Method

get_moments(
    𝓂,
    parameters;
    non_stochastic_steady_state,
    standard_deviation,
    variance,
    covariance
)

Return the first and second moments of endogenous variables using the linearised solution. By default returns: non stochastic steady state (SS), and standard deviations, but can also return variances, and covariance matrix. Function to use when differentiating model moments with repect to parameters.

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

parameters [Type: Vector]: Parameter values in alphabetical order (sorted by parameter name).

Keyword Arguments

non_stochastic_steady_state [Default: true, Type: Bool]: switch to return SS of endogenous variables

standard_deviation [Default: true, Type: Bool]: switch to return standard deviation of endogenous variables

variance [Default: false, Type: Bool]: switch to return variance of endogenous variables

covariance [Default: false, Type: Bool]: switch to return covariance matrix of endogenous variables

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

get_moments(RBC, RBC.parameter_values)
# output
2-element Vector{Any}:
 [5.936252888048724, 47.39025414828808, 6.884057971014486, 0.0]
 [0.02666420378525522, 0.26467737291222343, 0.07393254045396495, 0.010206207261596576]

MacroModelling.get_moments — Method

get_moments(
    𝓂;
    parameters,
    non_stochastic_steady_state,
    standard_deviation,
    variance,
    covariance,
    derivatives,
    parameter_derivatives
)

Return the first and second moments of endogenous variables using the linearised solution. By default returns: non stochastic steady state (SS), and standard deviations, but can also return variances, and covariance matrix. Function to use when differentiating model moments with repect to parameters.

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

Keyword Arguments

parameters : If nothing is provided, the solution is calculated for the parameters defined previously. If a vector with parameter values, or a named tuple is provided and the parameters differ from the previously defined the solution will be recalculated.

non_stochastic_steady_state [Default: true, Type: Bool]: switch to return SS of endogenous variables

standard_deviation [Default: true, Type: Bool]: switch to return standard deviation of endogenous variables

variance [Default: false, Type: Bool]: switch to return variance of endogenous variables

covariance [Default: false, Type: Bool]: switch to return covariance matrix of endogenous variables

derivatives [Default: true, Type: Bool]: switch to calculate derivatives of SS, standard deviation, and variance with respect to the parameters

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

moments = get_moments(RBC);

moments[1]
# output
2-dimensional KeyedArray(NamedDimsArray(...)) with keys:
↓   Variables ∈ 4-element Vector{Symbol}
→   Steady_state_and_∂steady_state∂parameter ∈ 6-element Vector{Symbol}
And data, 4×6 Matrix{Float64}:
        (:Steady_state)  (:std_z)  (:ρ)     (:δ)      (:α)       (:β)
  (:c)   5.93625          0.0       0.0   -116.072    55.786     76.1014
  (:k)  47.3903           0.0       0.0  -1304.95    555.264   1445.93
  (:q)   6.88406          0.0       0.0    -94.7805   66.8912   105.02
  (:z)   0.0              0.0       0.0      0.0       0.0        0.0

moments[2]
# output
2-dimensional KeyedArray(NamedDimsArray(...)) with keys:
↓   Variables ∈ 4-element Vector{Symbol}
→   Standard_deviation_and_∂standard_deviation∂parameter ∈ 6-element Vector{Symbol}
And data, 4×6 Matrix{Float64}:
        (:Standard_deviation)  (:std_z)  …  (:δ)       (:α)       (:β)
  (:c)   0.0266642              2.66642     -0.384359   0.2626     0.144789
  (:k)   0.264677              26.4677      -5.74194    2.99332    6.30323
  (:q)   0.0739325              7.39325     -0.974722   0.726551   1.08
  (:z)   0.0102062              1.02062      0.0        0.0        0.0

MacroModelling.get_non_stochastic_steady_state — Method

Wrapper for get_steady_state with stochastic = false.

MacroModelling.get_perturbation_solution — Function

See get_solution

MacroModelling.get_solution — Method

get_solution(𝓂; parameters)

Return the linearised solution and the non stochastic steady state (SS) of the model.

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

Keyword Arguments

parameters : If nothing is provided, the solution is calculated for the parameters defined previously. If a vector with parameter values, or a named tuple is provided and the parameters differ from the previously defined the solution will be recalculated.

The returned KeyedArray shows the SS, policy and transition functions of the model. The columns show the varibales including auxilliary endogenous and exogenous variables (due to leads and lags > 1). The rows are the SS, followed by the states, and exogenous shocks. Subscripts following variable names indicate the timing (e.g. variable₍₋₁₎ indicates the variable being in the past). Superscripts indicate leads or lags (e.g. variableᴸ⁽²⁾ indicates the variable being in lead by two periods). If no super- or subscripts follow the variable name, the variable is in the present.

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

get_solution(RBC)
# output
2-dimensional KeyedArray(NamedDimsArray(...)) with keys:
↓   Steady_state__States__Shocks ∈ 4-element Vector{Symbol}
→   Variable ∈ 4-element Vector{Symbol}
And data, 4×4 adjoint(::Matrix{Float64}) with eltype Float64:
                   (:c)         (:k)        (:q)        (:z)
  (:Steady_state)   5.93625     47.3903      6.88406     0.0
  (:k₍₋₁₎)          0.0957964    0.956835    0.0726316  -0.0
  (:z₍₋₁₎)          0.134937     1.24187     1.37681     0.2
  (:eps_z₍ₓ₎)       0.00674687   0.0620937   0.0688406   0.01

MacroModelling.get_steady_state — Method

get_steady_state(
    𝓂;
    parameters,
    derivatives,
    stochastic,
    parameter_derivatives
)

Return the (non stochastic) steady state and derivatives with respect to model parameters.

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

Keyword Arguments

parameters : If nothing is provided, the solution is calculated for the parameters defined previously. If a vector with parameter values, or a named tuple is provided and the parameters differ from the previously defined the solution will be recalculated.

derivatives [Default: true, Type: Bool]: calculate derivatives of the SS with respect to the parameters

stochastic [Default: false, Type: Bool]: return stochastic steady state using second order perturbation. No derivatives are calculated.

The columns show the SS and parameters for which derivatives are taken. The rows show the variables.

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

get_steady_state(RBC)
# output
2-dimensional KeyedArray(NamedDimsArray(...)) with keys:
↓   Variables_and_calibrated_parameters ∈ 4-element Vector{Symbol}
→   Steady_state_and_∂steady_state∂parameter ∈ 6-element Vector{Symbol}
And data, 4×6 Matrix{Float64}:
        (:Steady_state)  (:std_z)  (:ρ)     (:δ)      (:α)       (:β)
  (:c)   5.93625          0.0       0.0   -116.072    55.786     76.1014
  (:k)  47.3903           0.0       0.0  -1304.95    555.264   1445.93
  (:q)   6.88406          0.0       0.0    -94.7805   66.8912   105.02
  (:z)   0.0              0.0       0.0      0.0       0.0        0.0

MacroModelling.get_stochastic_steady_state — Method

Wrapper for get_steady_state with stochastic = true.

MacroModelling.plot — Method

plot(
    𝓂;
    periods,
    shocks,
    variables,
    parameters,
    show_plots,
    save_plots,
    save_plots_format,
    save_plots_path,
    plots_per_page,
    algorithm,
    negative_shock,
    generalised_irf,
    initial_state
)

Plot impulse response functions (IRFs) of the model.

The left axis shows the level, and the right the deviation from the reference steady state. Linear solutions have the non stochastic steady state as reference other solutoin the stochastic steady state. The horizontal black line indicates the reference steady state. Variable names are above the subplots and the title provides information about the model, shocks and number of pages per shock.

Arguments

𝓂: the object created by @model and @parameters for which to get the solution.

Keyword Arguments

plots_per_page [Default: 9, Type: Int]: how many plots to show per page

save_plots [Default: false, Type: Bool]: switch to save plots using path and extension from save_plots_path and save_plots_format. Separate files per shocks and variables depending on number of variables and plots_per_page

save_plots_path [Default: pwd(), Type: String]: path where to save plots

save_plots_format [Default: :pdf, Type: Symbol]: output format of saved plots. See input formats compatible with GR for valid formats.

show_plots [Default: true, Type: Bool]: show plots. Separate plots per shocks and varibles depending on number of variables and plots_per_page.

periods [Default: 40, Type: Int]: number of periods for which to calculate the IRFs

algorithm [Default: :first_order, Type: Symbol]: solution algorithm for which to show the IRFs

parameters : If nothing is provided, the solution is calculated for the parameters defined previously. If a vector with parameter values, or a named tuple is provided and the parameters differ from the previously defined the solution will be recalculated.

variables [Default: :all]: variables for which to calculate the IRFs. Inputs can be either a Symbol (e.g. :y or :all), Matrix{Symbol} or Vector{Symbol}. Any variables not part of the model will trigger a warning.

shocks [Default: :all]: shocks for which to calculate the IRFs. Inputs can be either a Symbol (e.g. :y, :simulate, :none, or :all), Matrix{Symbol} or Vector{Symbol}. :simulate triggers random draws of all shocks. Any shocks not part of the model will trigger a warning. :none in combination with an initial_state can be used for deterministic simulations.

negative_shock [Default: false, Type: Bool]: calculate a negative shock. Relevant for generalised IRFs.

generalised_irf [Default: false, Type: Bool]: calculate generalised IRFs. Relevant for nonlinear solutions.

initial_state [Default: [0.0], Type: Vector{Float64}]: provide state from which to start IRFs. Relevant for normal IRFs.

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end;

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end;

plot(RBC)

MacroModelling.plot_IRF — Function

See plot

MacroModelling.plot_irf — Function

See plot

MacroModelling.plot_irfs — Function

See plot

MacroModelling.plot_simulations — Method

Wrapper for plot with shocks = :simulate and periods = 100.

MacroModelling.simulate — Method

Wrapper for get_irf with shocks = :simulate.

MacroModelling.@model — Macro

Parses the model equations and assigns them to an object.

Arguments

𝓂: name of the object to be created containing the model information.

ex: equations

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end

MacroModelling.@parameters — Macro

Adds parameter values and calibration equations to the previously defined model.

Arguments

𝓂: name of the object previously created containing the model information.

ex: parameter, parameters values, and calibration equations

Examples

using MacroModelling

@model RBC begin
    1  /  c[0] = (β  /  c[1]) * (α * exp(z[1]) * k[0]^(α - 1) + (1 - δ))
    c[0] + k[0] = (1 - δ) * k[-1] + q[0]
    q[0] = exp(z[0]) * k[-1]^α
    z[0] = ρ * z[-1] + std_z * eps_z[x]
end

@parameters RBC begin
    std_z = 0.01
    ρ = 0.2
    δ = 0.02
    α = 0.5
    β = 0.95
end